Gemma4-small ViT review: consolidate clip tests into gemma4_small_layers_test.py; optimizer clip-freeze comment; E2B multimodal e2e script
Addresses hengtaoguo review on #4790:
- move gemma4_clip_linears_test.py + gemma4_mm_contract_gate_test.py into
  tests/unit/gemma4_small_layers_test.py (26 methods; fixes a stray mid-file
  main-guard that shadowed OptionSInvariantsTest; fixes unused-var + adds docstring)
- optimizers.py: one-line callout for the Gemma4-small ViT clip-bound freeze
- add tests/end_to_end/tpu/gemma4/e2b/test_gemma4_multimodal.sh (image e2e on the NNX path)
